chore: split into components
